% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wildfinder_data.R
\name{dl_wildfinder_data}
\alias{dl_wildfinder_data}
\title{Download WILDFINDER dataset}
\usage{
dl_wildfinder_data(x)
}
\value{
This function returns the path (`character`) to the downloaded file.
}
\description{
This function downloads the WINDERFINDER dataset (csv file) hosted on the 
GitHub repository <https://github.com/rdatatoolbox/datarepo/>.
The file starting with `wildfinder` will be stored in 
`data/wildfinder/`. This folder will be created if required.
}
